Make plotting backends extensions
